[update] 메인페이지 리펙토링 #72
